refactor: move context out of http-request module
