+++ src/http-request.ts
@@ -15,18 +15,13 @@
  */
 
 /*
- * This file contains http helper functions, used in both middleware and
- * non-middleware scenarios, to extract or inject useful LogEntry data to/fro
- * HTTP headers.
+ * This file contains helpers, used in both middleware and non-middleware
+ * scenarios, to format http.incomingMessage into structured HTTPRequests.
  *
- * Example:
- * - If header context is available from OpenCensus or OpenTelemetry (via W3
- * context propagation protocol), extract a log trace field.
- * - If context is available from X-Cloud-Trace-Context header, extract trace,
- * spanId, and traceSampled fields.
+ * See more:
+ * https://cloud.google.com/logging/docs/reference/v2/rest/v2/LogEntry#httprequest
  */
 
-import * as w3cContext from '@opencensus/propagation-stackdriver';
 import * as http from 'http';
 export interface CloudLoggingHttpRequest {
   requestMethod?: string;
@@ -55,8 +50,7 @@ export interface ServerRequest extends http.IncomingMessage {
 
 /**
  * makeHttpRequestData turns raw incoming HTTPRequests into structured
- * HTTPRequest objects interpreted by Cloud Logging. See more:
- * https://cloud.google.com/logging/docs/reference/v2/rest/v2/LogEntry#httprequest
+ * HTTPRequest objects interpreted by Cloud Logging.
  *
  * @param req
  * @param res
@@ -117,122 +111,4 @@ export function makeHttpRequestData(
     status ? {status} : null,
     latency ? {latency} : null
   );
-}
-
-/**
- * HeaderWrapper: wraps getHeader and setHeader operations.
- */
-export type HeaderWrapper = w3cContext.HeaderGetter & w3cContext.HeaderSetter;
-
-/**
- * makeHeaderWrapper returns a wrapper with set and get header functionality,
- * returning null if the incoming request object doesn't contain the 'header'
- * propery.
- * @param req
- */
-export function makeHeaderWrapper(
-  req: http.IncomingMessage
-): HeaderWrapper | null {
-  if (!req.headers) return null;
-  const wrapper = {
-    setHeader(name: string, value: string) {
-      req.headers[name] = value;
-    },
-    getHeader(name: string) {
-      return req.headers[name];
-    },
-  };
-  return wrapper;
-}
-
-export interface CloudTraceContext {
-  trace: string;
-  spanId?: string;
-  traceSampled?: boolean;
-}
-
-/**
- * getTraceContext returns a CloudTraceContext with as many available trace and
- * span properties as possible. It examines HTTP headers for trace context.
- * Optionally, it can inject a W3C compliant trace context when no context is
- * available from headers.
- */
-export function getTraceContext(
-  req: http.IncomingMessage,
-  projectId: string,
-  inject?: boolean
-): CloudTraceContext {
-  const context: CloudTraceContext = {
-    trace: '',
-    spanId: undefined,
-    traceSampled: undefined,
-  };
-  const wrapper = makeHeaderWrapper(req);
-  if (wrapper) {
-    // Detect 'X-Cloud-Trace-Context' header.
-    const cloudContext = getCloudTraceContext(wrapper);
-    if (cloudContext) {
-      context.trace = `projects/${projectId}/traces/${cloudContext.trace}`;
-      context.spanId = cloudContext.spanId;
-      context.traceSampled = cloudContext.traceSampled;
-    } else {
-      // Detect 'traceparent' header, optionally injecting one if not available.
-      const parentContext = getOrInjectTraceParent(wrapper, inject);
-      if (parentContext) {
-        context.trace = `projects/${projectId}/traces/${parentContext.traceId}`;
-        context.spanId = parentContext.spanId;
-      }
-    }
-  }
-  return context;
-}
-
-/**
- * getCloudTraceContext looks for the HTTP header 'X-Cloud-Trace-Context'
- * per Cloud Trace specifications. `traceSampled` is false by default.
- * @param headerWrapper
- */
-export function getCloudTraceContext(
-  headerWrapper: HeaderWrapper
-): CloudTraceContext | null {
-  // Infer trace & span if not user specified already
-  const regex = /([a-f\d]+)?(\/?([a-f\d]+))?(;?o=(\d))?/;
-  const match = headerWrapper
-    .getHeader('x-cloud-trace-context')
-    ?.toString()
-    .match(regex);
-  if (match) {
-    return {
-      trace: match[1],
-      spanId: match[3],
-      traceSampled: match[5] === '1',
-    };
-  } else {
-    return null;
-  }
-}
-
-/**
- * getCloudTraceContext looks for the HTTP header 'traceparent'
- * per W3C specifications for OpenTelemetry and OpenCensus
- * Read more about W3C protocol: https://www.w3.org/TR/trace-context/
- *
- * Optional: Inject a W3C compliant context with trace and span in the header,
- * if not available.
- *
- * @param headerWrapper
- * @param inject
- */
-export function getOrInjectTraceParent(
-  headerWrapper: HeaderWrapper,
-  inject?: boolean
-): w3cContext.SpanContext | null {
-  let spanContext = w3cContext.extract(headerWrapper);
-  if (spanContext) return spanContext;
-  if (inject) {
-    // We were the first actor to detect lack of context. Establish context.
-    spanContext = w3cContext.generate();
-    w3cContext.inject(headerWrapper, spanContext);
-  }
-  return spanContext;
 }
